引入Zap框架处理日志，放弃GORM转用sqlx实现数据库操作

+++ serializer/user_points.go
@@ -4,9 +4,9 @@ import "acgfate/model"
 
 // UserPointsResponse 用户点数信息
 type UserPointsResponse struct {
-	EXP   int64 `json:"exp"`
-	Level uint8 `json:"level"`
-	Coins uint  `json:"coins"`
+	EXP   uint16 `json:"exp"`
+	Level uint8  `json:"level"`
+	Coins uint64 `json:"coins"`
 }
 
 // BuildUserPointsResponse 用户点数信息返回构建
